lead-scraper-v5/requirements.txt
# Core scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Browser automation with stealth
undetected-chromedriver>=3.5.5
selenium>=4.15.0
selenium-wire>=5.1.0
webdriver-manager>=4.0.0

# Data processing
pandas>=2.1.0
openpyxl>=3.1.0
numpy>=1.24.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Utilities
validators>=0.22.0
tqdm>=4.66.0
colorama>=0.4.6

# Optional: Proxy support
# requests[socks]>=2.31.0
# pysocks>=1.7.1